+++ commands/System/help.js
@@ -33,7 +33,7 @@ exports.run = (client, msg, [cmd]) => {
 
 exports.conf = {
   enabled: true,
-  guildOnly: false,
+  runIn: ["text", "dm", "group"],
   aliases: [],
   permLevel: 0,
   botPerms: [],