+++ commands/System/reboot.js
@@ -10,7 +10,7 @@ exports.run = (client, msg) => {
 
 exports.conf = {
   enabled: true,
-  guildOnly: false,
+  runIn: ["text", "dm", "group"],
   aliases: [],
   permLevel: 10,
   botPerms: [],